+++ rflx/const.py
@@ -1,3 +1,4 @@
+import multiprocessing
 from pathlib import Path
 from typing import Final
 
@@ -10,3 +11,7 @@ INTERNAL_PACKAGE: Final = ID("__INTERNAL__")
 MAX_SCALAR_SIZE: Final = 63
 
 CACHE_PATH: Final = Path.home() / ".cache" / "RecordFlux"
+
+# The use of "forkserver" or "spawn" as start method for starting processes prevents deadlocks when
+# RecordFlux is executed by another process, e.g., when the language server is started by VS Code.
+MP_CONTEXT = multiprocessing.get_context("forkserver")